mail_collector/requirements.txt
configparser
flask
MySQL-python
mysqlclient
python-dateutil
Wekzeug
